reconfirm: Add support to avahi core
Adds support to avahi core for reconfirming a record, as
described in RFC 6762 section 10.4. Four states were added
to the cache entry to manage re-querying the record. If
the service is no longer available then four separate
queries are issued at an interval of 1.5 seconds. If a
response is received before the final state then the
states ends up being reset to AVAHI_CACHE_VALID. If a
response is not received within 6 seconds then the entry
is removed from the cache.

Renames expire_in_one_second() to update_state_set_timer()
to allow for different state timeouts.

Signed-off-by: Nate Karstens <nate.karstens@garmin.com>